+++ tools/ckan-utils/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 14.January.2024
 - adds `User-Agent` header to `opentransportdata.swiss` API requests
+- adds `fetch_metadata_cli.py` script that fetches JSON metadata from the CKAN API
 
 23.September.2024
 - use https://opentransportdata.swiss/de/dataset/go-realtime dataset instead of static CSV export